GEOMESA-509 Factor out useful methods out of IndexSchema
* split off sensible things from IndexSchema into a trait.
