+++ geomesa-core/src/main/scala/org/locationtech/geomesa/core/index/IndexSchema.scala
@@ -17,18 +17,12 @@
 package org.locationtech.geomesa.core.index
 
 import com.typesafe.scalalogging.slf4j.Logging
-import com.vividsolutions.jts.geom.{Geometry, GeometryCollection, Point, Polygon}
+import com.vividsolutions.jts.geom.Point
 import org.apache.accumulo.core.data.Key
 import org.geotools.data.Query
-import org.joda.time.format.DateTimeFormat
-import org.joda.time.{DateTime, DateTimeZone, Interval}
 import org.locationtech.geomesa.core.data._
 import org.locationtech.geomesa.core.util._
-import org.locationtech.geomesa.utils.text.WKTUtils
 import org.opengis.feature.simple.{SimpleFeature, SimpleFeatureType}
-
-import scala.annotation.tailrec
-import scala.util.parsing.combinator.RegexParsers
 
 // A secondary index consists of interleaved elements of a composite key stored in
 // Accumulo's key (row, column family, and column qualifier)
@@ -88,85 +82,7 @@ case class IndexSchema(encoder: IndexEntryEncoder,
   }
 }
 
-object IndexSchema extends RegexParsers with Logging {
-  val minDateTime = new DateTime(0, 1, 1, 0, 0, 0, DateTimeZone.forID("UTC"))
-  val maxDateTime = new DateTime(9999, 12, 31, 23, 59, 59, DateTimeZone.forID("UTC"))
-  val everywhen = new Interval(minDateTime, maxDateTime)
-  val everywhere = WKTUtils.read("POLYGON((-180 -90, 0 -90, 180 -90, 180 90, 0 90, -180 90, -180 -90))").asInstanceOf[Polygon]
-
-  def somewhen(interval: Interval): Option[Interval] =
-    interval match {
-      case null                => None
-      case i if i == everywhen => None
-      case _                   => Some(interval)
-    }
-
-  def innerSomewhere(geom: Geometry): Option[Geometry] =
-    geom match {
-      case null                 => None
-      case p if p == everywhere => None
-      case g: Geometry          => Some(g)
-      case _                    => None
-    }
-
-  // This function helps catch nulls and 'entire world' polygons.
-  def somewhere(geom: Geometry): Option[Geometry] =
-    geom match {
-      case null => None
-      case gc: GeometryCollection =>
-        val wholeWorld = (0 until gc.getNumGeometries).foldRight(false) {
-          case (i, seenEverywhere) => gc.getGeometryN(i).equals(everywhere) || seenEverywhere
-        }
-        if(wholeWorld) None else Some(gc)
-      case g: Geometry => innerSomewhere(g)
-    }
-
-  val DEFAULT_TIME = new DateTime(0, DateTimeZone.forID("UTC"))
-
-  val CODE_START = "%"
-  val CODE_END = "#"
-  val GEO_HASH_CODE = "gh"
-  val DATE_CODE = "d"
-  val CONSTANT_CODE = "cstr"
-  val RANDOM_CODE = "r"
-  val SEPARATOR_CODE = "s"
-  val ID_CODE = "id"
-  val PART_DELIMITER = "::"
-
-  def pattern[T](p: => Parser[T], code: String): Parser[T] = CODE_START ~> p <~ (CODE_END + code)
-
-  // A separator character, typically '%~#s' would indicate that elements are to be separated
-  // with a '~'
-  def sep = pattern("\\W".r, SEPARATOR_CODE)
-
-  // A random partitioner.  '%999#r' would write a random value between 000 and 999 inclusive
-  def randPartitionPattern = pattern("\\d+".r,RANDOM_CODE)
-  def randEncoder: Parser[PartitionTextFormatter] = randPartitionPattern ^^ {
-    case d => PartitionTextFormatter(d.toInt)
-  }
-
-  def offset = "[0-9]+".r ^^ { _.toInt }
-  def bits = "[0-9]+".r ^^ { _.toInt }
-
-  // A geohash encoder.  '%2,4#gh' indicates that two characters starting at character 4 should
-  // be extracted from the geohash and written to the field
-  def geohashPattern = pattern((offset <~ ",") ~ bits, GEO_HASH_CODE)
-  def geohashEncoder: Parser[GeoHashTextFormatter] = geohashPattern ^^ {
-    case o ~ b => GeoHashTextFormatter(o, b)
-  }
-
-  // A date encoder. '%YYYY#d' would pull out the year from the date and write it to the key
-  def datePattern = pattern("\\w+".r, DATE_CODE)
-  def dateEncoder: Parser[DateTextFormatter] = datePattern ^^ {
-    case t => DateTextFormatter(t)
-  }
-
-  // A constant string encoder. '%fname#cstr' would yield fname
-  //  We match any string other that does *not* contain % or # since we use those for delimiters
-  def constStringPattern = pattern("[^%#]+".r, CONSTANT_CODE)
-  def constantStringEncoder: Parser[ConstantTextFormatter] = constStringPattern ^^ {
-    case str => ConstantTextFormatter(str)
-  }
+object IndexSchema extends SchemaHelpers with Logging {
 
   // a key element consists of a separator and any number of random partitions, geohashes, and dates
   def keypart: Parser[CompositeTextFormatter] =
@@ -190,15 +106,6 @@ object IndexSchema extends RegexParsers with Logging {
     val (rowf, cff, cqf) = parse(formatter, s).get
     IndexEntryEncoder(rowf, cff, cqf, featureEncoder)
   }
-
-  // extracts an entire date encoder from a key part
-  @tailrec
-  def extractDateEncoder(seq: Seq[TextFormatter], offset: Int, sepLength: Int): Option[(String, Int)] =
-    seq match {
-      case DateTextFormatter(f)::xs => Some(f,offset)
-      case x::xs => extractDateEncoder(xs, offset + x.numBits + sepLength, sepLength)
-      case Nil => None
-    }
 
   // builds the date decoder to deserialize the entire date from the parts of the index key
   def dateDecoderParser = keypart ~ PART_DELIMITER ~ keypart ~ PART_DELIMITER ~ cqpart ^^ {
@@ -231,15 +138,6 @@ object IndexSchema extends RegexParsers with Logging {
 
   def buildDateDecoder(s: String): Option[DateDecoder] = parse(dateDecoderParser, s).get
 
-  // extracts the geohash encoder from a keypart
-  @tailrec
-  def extractGeohashEncoder(seq: Seq[TextFormatter], offset: Int, sepLength: Int): (Int, (Int, Int)) =
-    seq match {
-      case GeoHashTextFormatter(off, bits)::xs => (offset, (off, bits))
-      case x::xs => extractGeohashEncoder(xs, offset + x.numBits + sepLength, sepLength)
-      case Nil => (0,(0,0))
-    }
-
   // builds a geohash decoder to extract the entire geohash from the parts of the index key
   def ghDecoderParser = keypart ~ PART_DELIMITER ~ keypart ~ PART_DELIMITER ~ cqpart ^^ {
     case rowf ~ PART_DELIMITER ~ cff ~ PART_DELIMITER ~ cqf => {
@@ -255,18 +153,6 @@ object IndexSchema extends RegexParsers with Logging {
 
   def buildGeohashDecoder(s: String): GeohashDecoder = parse(ghDecoderParser, s).get
 
-  def extractIdEncoder(seq: Seq[TextFormatter], offset: Int, sepLength: Int): Int =
-    seq match {
-      case IdFormatter(maxLength)::xs => maxLength
-      case _ => sys.error("Id must be first element of column qualifier")
-    }
-
-  // An id encoder. '%15#id' would pad the id out to 15 characters
-  def idEncoder: Parser[IdFormatter] = pattern("[0-9]*".r, ID_CODE) ^^ {
-    case len if len.length > 0 => IdFormatter(len.toInt)
-    case _                     => IdFormatter(0)
-  }
-
   def idDecoderParser = keypart ~ PART_DELIMITER ~ keypart ~ PART_DELIMITER ~ cqpart ^^ {
     case rowf ~ PART_DELIMITER ~ cff ~ PART_DELIMITER ~ cqf => {
       val bits = extractIdEncoder(cqf.lf, 0, cqf.sep.length)
@@ -275,18 +161,6 @@ object IndexSchema extends RegexParsers with Logging {
   }
 
   def buildIdDecoder(s: String) = parse(idDecoderParser, s).get
-
-  def constStringPlanner: Parser[ConstStringPlanner] = constStringPattern ^^ {
-    case str => ConstStringPlanner(str)
-  }
-
-  def randPartitionPlanner: Parser[RandomPartitionPlanner] = randPartitionPattern ^^ {
-    case d => RandomPartitionPlanner(d.toInt)
-  }
-
-  def datePlanner: Parser[DatePlanner] = datePattern ^^ {
-    case fmt => DatePlanner(DateTimeFormat.forPattern(fmt))
-  }
 
   def geohashKeyPlanner: Parser[GeoHashKeyPlanner] = geohashPattern ^^ {
     case o ~ b => GeoHashKeyPlanner(o, b)
@@ -301,7 +175,6 @@ object IndexSchema extends RegexParsers with Logging {
     case Success(result, _) => result
     case fail: NoSuccess => throw new Exception(fail.msg)
   }
-
 
   def geohashColumnFamilyPlanner: Parser[GeoHashColumnFamilyPlanner] = (keypart ~ PART_DELIMITER) ~> (sep ~ rep(randEncoder | geohashEncoder | dateEncoder | constantStringEncoder)) <~ (PART_DELIMITER ~ keypart) ^^ {
     case sep ~ xs => xs.find(tf => tf match {
